+++ skbio/stats/distance/_base.py
@@ -21,6 +21,8 @@ from skbio.util import find_duplicates
 from skbio.util._decorator import experimental, classonlymethod
 from skbio.util._misc import resolve_key
 
+from ._utils import distmat_reorder, distmat_reorder_condensed
+
 
 class DissimilarityMatrixError(Exception):
     """General error for dissimilarity matrix validation failures."""
@@ -387,7 +389,8 @@ class DissimilarityMatrix(SkbioObject):
                     pass
             ids = found_ids
 
-        filtered_data = self._data[idxs][:, idxs]
+        filtered_data = distmat_reorder(self._data, idxs)
+
         return self.__class__(filtered_data, ids)
 
     def _stable_order(self, ids):
@@ -1063,11 +1066,12 @@ class DistanceMatrix(DissimilarityMatrix):
 
         """
         order = np.random.permutation(self.shape[0])
-        permuted = self._data[order][:, order]
 
         if condensed:
-            return squareform(permuted, force='tovector', checks=False)
+            permuted_condensed = distmat_reorder_condensed(self._data, order)
+            return permuted_condensed
         else:
+            permuted = distmat_reorder(self._data, order)
             return self.__class__(permuted, self.ids)
 
     def _validate(self, data, ids):